Add Entity Framework and DB context to MVC project
In this update, we've integrated Microsoft's Entity Framework into our MVC project. We've also added a new ApplicationDbContext to the service container, which will be using SQL Server with a connection string named "DefaultConnection". This allows us to interact with our database more efficiently and in a more structured manner.