Alter DisplayOrder type in Categories table
This commit introduces a change in the data type of the 'DisplayOrder' field in the 'Categories' table. The previous string type has been replaced with an integer type to better reflect its purpose and usage. This change is reflected in both the migration file and model snapshot. The Down method is also updated to revert this change if needed.