Add category creation functionality
A new feature has been added to the CategoryController. This includes a Create() method that displays a view for creating a new category, and another Create() method that accepts a POST request with the new category object. After adding the new category to the database, it redirects back to the index page.