Add success messages and delete functionality to CategoryController
In this update, we've added success messages for creating and updating categories in the CategoryController. Now, users will receive a confirmation message after successfully performing these actions. Additionally, we've implemented a delete function for categories. This includes both the GET method to display the category to be deleted and the POST method to perform the deletion. After successful deletion, a success message is displayed as well.