Add Categories Index page
A new Razor Page for displaying a list of categories has been added. This page uses the ApplicationDbContext to fetch all categories from the database and display them on the page. The IndexModel class, which is part of this page, contains a constructor that takes an instance of ApplicationDbContext as a parameter and assigns it to a private field. It also includes an OnGet method that populates the CategoryList property with data fetched from the database.