Update default routing pattern with area
The default routing pattern in the MVC application has been updated. Previously, it was set to "{controller=Home}/{action=Index}/{id?}", but now it includes an area parameter and is changed to "{area=Customer}/{controller=Home}/{action=Index}/{id?}". This change allows for better organization of controllers and views based on their functional areas.