Add product deletion view
A new view has been added to the admin area for deleting products. This view includes a form with fields for various product attributes, all of which are disabled and only serve to display the current values of the product being deleted. The form also includes a 'Delete' button to submit the form and delete the product, as well as a 'Back' button that redirects back to the product index page. Validation scripts have also been included at the end of this file.