Implement automatic database migration
Added functionality to automatically migrate the database if there are any pending migrations. This is achieved by creating a scope at startup, getting the required service for ApplicationDbContext, and then checking for any pending migrations. If there are any, the database is migrated accordingly. Also added Microsoft.Extensions.Hosting to the using directives.