Add foreign key for Category-Product relation
This commit introduces a new foreign key to establish a relationship between the 'Category' and 'Product' entities. The 'CategoryId' column has been added to the 'Products' table, and corresponding updates have been made to existing data. An index on this new column has also been created for performance optimization. In case of rollback, the changes can be safely reverted by dropping the foreign key, index, and column.