Add ImageUrl field to Products table
A new migration has been added that introduces an 'ImageUrl' field to the 'Products' table. This is a nullable string type column. The existing data in the 'Products' table has been updated to set this new column's value as null for all current entries. In case of rollback, the migration will remove the 'ImageUrl' column from the 'Products' table. The model snapshot has also been updated to reflect this change in schema.