Add image handling to ProductController
The ProductController has been updated to handle image uploads. The constructor now takes an additional IWebHostEnvironment parameter, which is used to get the root path of the web host. When a new product is added, if an image file is provided, it's saved in the 'images\product' directory under the web root path with a unique filename. The URL of this image is then stored in the product's ImageUrl property before updating it in the database.