Refactor product table and add script
Simplified the product table in the admin view by removing several columns and adding an id for easier manipulation. Removed the loop that was generating rows from the model, as this will now be handled by a new JavaScript file. Added a reference to this new 'product.js' script at the end of the file.