Integrate IUnitOfWork into HomeController
The HomeController now uses the IUnitOfWork interface for data access. This change allows us to fetch and display a list of products on the Index view, as well as individual product details on a new Details view. The constructor has been updated to include an instance of IUnitOfWork, which is then used in the Index and Details actions.