Add role selection to user registration
Added a new feature that allows users to select their role during the registration process. This was achieved by adding a 'Role' property and an 'IEnumerable<SelectListItem> RoleList' property in the InputModel class. The 'RoleList' is populated with roles from the RoleManager during GET requests, and the selected role is assigned to the user during POST requests. If no role is selected, the user will be assigned a default customer role.