Add Faction List view in Admin area
A new view has been added to the Admin area for managing Factions. This includes a table layout for displaying faction details such as name, address, city, state, postcode and phone number. Also included is a button for creating new factions. The associated script file 'faction.js' is linked at the end of the file.