Add Faction Upsert view
A new view for the Faction model has been added in the Admin area. This view, named 'Upsert', allows both creation and updating of Faction instances. It includes form fields for all properties of the Faction model, along with validation summaries for each field. The TinyMCE editor is also initialized in this view to enhance text input capabilities.