Add FactionId to AspNetUsers table
This commit introduces a new column, FactionId, to the AspNetUsers table. This change includes creating an index for the new column and setting up a foreign key relationship with the Factions table. The migration files have been updated accordingly. In addition, the ApplicationDbContextModelSnapshot has been modified to reflect these changes in the model structure.