Add Faction support to user registration
The user registration process has been updated to include support for Factions. A new dependency, IUnitOfWork, was injected into the RegisterModel constructor and used to populate a list of available factions. This list is then displayed in the registration form if the user selects a role associated with a faction. The selected faction ID is stored in the User object upon successful registration.