Add faction selection to registration page
A new dropdown for faction selection has been added to the registration page. This dropdown is hidden by default and only becomes visible when 'Faction' is selected in the role dropdown. The visibility of this new dropdown is controlled using a jQuery script that listens for changes on the role dropdown.