Update Details view to use ShoppingCart model
The Details view in the Customer area has been updated to use the ShoppingCart model instead of the Product model. This change includes updates to all references from 'Model' to 'Model.Product'. Two hidden input fields for ProductId and Id have been added. The "Add to Cart" button is now enabled, and an input field for Count has been introduced.