Add shopping cart view
A new view has been added for the shopping cart. This includes a list of items in the cart, with options to increase or decrease quantity, remove items, and continue shopping. The total price is also displayed at the bottom along with a button to proceed to the summary page.