Add SessionId and PaymentIntentId to OrderHeader
This commit introduces two new columns, 'SessionId' and 'PaymentIntentId', to the 'OrderHeaders' table. These fields are nullable and of type nvarchar(max). The migration file for adding these columns has been created along with the corresponding down migration for removing them if needed. The ApplicationDbContextModelSnapshot has also been updated to reflect these changes in the model.