Add payment status check in OrderConfirmation
Added a new feature to the OrderConfirmation method in the CartController. Now, it checks if the order's payment status is not delayed. If so, it retrieves the session details and updates the StripePaymentID and Status of the order if the session's PaymentStatus is 'paid'. After these operations, all shopping cart items associated with that user are deleted.